IVC integration: reorganize + fix compilation of test_simple_add
